# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Project Alpha API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      description: Returns a presigned GCS URL for direct file upload. The client uploads the file directly to GCS using the returned URL.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RequestUploadUrlBody:
      type: object
      properties:
        name:
          type: string
          description: Original file name
        size:
          type: integer
          description: File size in bytes
        contentType:
          type: string
          description: MIME type of the file
      required:
        - name
        - size
        - contentType
    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
          description: Presigned GCS URL for direct upload
        objectPath:
          type: string
          description: Normalized object path to store in DB (e.g. /objects/uploads/uuid)
        metadata:
          type: object
          properties:
            name:
              type: string
            size:
              type: integer
            contentType:
              type: string
      required:
        - uploadURL
        - objectPath
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
        - error